[feat]: add SparkIngestionProperties ctor for IngestionProperties, move ingestion utils to KustoIngestionUtils object, fix dynamic csvMapping generation logic
